DYN-5674-CustomNode-Renamed (#13854)
After the auto-save functionality is triggered when we are located in the Custom Node edit tab when the node in the Home Workspace is being renamed to "backup". Then after my fix first is checking if it has a name and file associated (that usually happen when you create/edit a custom node) if that is the case then we don't execute the rename process.
Also I added a test that will validate that the Save functionality doesn't rename the CustomNode name.